Added process tools class
